Moved container to utils
